+++ src/api_handlers/json_models/chat_completion.rs
