+++ src/api_handlers/json_models/mod.rs
@@ -1,2 +1,2 @@
 pub mod chat_completion;
-pub mod language_models;
+pub mod language_model;
