+++ src/api_handlers/mod.rs
@@ -1,3 +1,4 @@
 mod api_instance;
 pub mod chat_completions;
-pub mod models;
+mod json_models;
+pub mod language_models;
